+++ lib/src/fx.dart
@@ -190,6 +190,8 @@ class Fx<T> extends Iterable<T> {
 
   T? maxBy(Object? Function(T a) f) => s.maxBy(f, _inner);
 
+  num sumBy(num Function(T a) f) => s.sumBy(f, _inner);
+
   (List<T>, List<T>) partition(bool Function(T a) f) => s.partition(f, _inner);
 
   int size() => s.size(_inner);
@@ -346,6 +348,8 @@ class FxAsync<T> implements FxAsyncIterable<T> {
   Future<T?> minBy(Object? Function(T a) f) => s.minByAsync(f, _inner);
 
   Future<T?> maxBy(Object? Function(T a) f) => s.maxByAsync(f, _inner);
+
+  Future<num> sumBy(FutureOr<num> Function(T a) f) => s.sumByAsync(f, _inner);
 
   Future<(List<T>, List<T>)> partition(FutureOr<bool> Function(T a) f) =>
       s.partitionAsync(f, _inner);